moved cli utils to utils
